Fix Conflict and adjust for npm package
- Update dev/index.ts example
- Move to src
- fix some types
- fix some props
